Cast int for number of partitions

+++ ooi_harvester/metadata/utils.py
@@ -360,7 +360,7 @@ def write_parquet(df, s3path):
 
 
 def df2parquet(df, table_name, bucket):
-    ddf = dd.from_pandas(df, npartitions=len(df) / 1000)
+    ddf = dd.from_pandas(df, npartitions=int(len(df) / 1000))
     s3path = os.path.join(bucket, table_name)
     write_parquet(ddf, s3path)
